# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Students",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000"
    }
  ],
  "paths": {
    "/api/students": {
      "get": {
        "tags": [
          "General"
        ],
        "summary": "Get all students",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/api/students/delete/{id}": {
      "delete": {
        "tags": [
          "General"
        ],
        "summary": "Delete student",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/api/students/add": {
      "post": {
        "tags": [
          "General"
        ],
        "summary": "Add student",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "name": "Mike",
                  "email": "mike@example.com",
                  "age": 12,
                  "cellphone": "12345"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/api/students/get/{id}": {
      "get": {
        "tags": [
          "General"
        ],
        "summary": "Get student",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "example": "7"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/api/students/change/{id}": {
      "put": {
        "tags": [
          "General"
        ],
        "summary": "Change student",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "age": 32,
                  "cellphone": "6789",
                  "email": "new_mike@example.com",
                  "name": "Mike Johns"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "example": "7"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/api/students/modify/{id}": {
      "patch": {
        "tags": [
          "General"
        ],
        "summary": "Modify student",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "age": 41
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "example": "7"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/api/health-check/ok": {
      "get": {
        "tags": [
          "General"
        ],
        "summary": "Health Check Ok",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/api/health-check/bad": {
      "get": {
        "tags": [
          "General"
        ],
        "summary": "Health Check Bad",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    }
  }
}